Update readme, place utilities in its own python file.
