+++ frontends/relay-futil/README.md
@@ -47,7 +47,7 @@ python3 example.py
 
 Pass the `-r` flag to this script to see the Relay code. Otherwise, we just print the FuTIL code. There is also an `-o` flag to try optimizing the Relay code a little bit.
 
-You can specify the name of an example as a command-line argument. The current options are `identity`, `add`, and `const`.
+You can specify the name of an example as a command-line argument. Currently, the only option is `identity`.
 
 
 Run the Tests
