+++ frontends/relay-futil/tests/add.relay
@@ -1,4 +1,4 @@
 v0.0.4
-fn (%x: int32) {
-  add(%x, 10)
+fn (%x: int32, %y: int32) {
+  add(%x, %y)
 }
